+++ app/src/main/java/com/empty/jinux/simplediary/ui/main/diarylist/DiaryListContract.kt
@@ -35,7 +35,7 @@ interface DiaryListContract {
 
         fun showAddDiary()
 
-        fun showDiaryDetailsUI(diaryId: Int)
+        fun showDiaryDetailsUI(diaryId: Long)
 
         fun showLoadingDiariesError()
